+++ lib/samtools-1.9/htslib-1.9/bgzf.c
@@ -2015,7 +2015,7 @@ int bgzf_index_load(BGZF *fp, const char *bname, const char *suffix)
     return -1;
 }
 
-int bgzf_useek(BGZF *fp, long uoffset, int where)
+int bgzf_useek(BGZF *fp, long long uoffset, int where)
 {
     if ( !fp->is_compressed )
     {
@@ -2072,7 +2072,7 @@ int bgzf_useek(BGZF *fp, long uoffset, int where)
     return 0;
 }
 
-long bgzf_utell(BGZF *fp)
+long long bgzf_utell(BGZF *fp)
 {
     return fp->uncompressed_address;    // currently maintained only when reading
 }
